Extract generator into a custom TimetableGridLayout
